+++ metricbeat/module/windows/service/zservice_windows.go
@@ -64,9 +64,9 @@ var (
 	procCloseServiceHandle    = modadvapi32.NewProc("CloseServiceHandle")
 )
 
-func _OpenSCManager(machineName *uint16, databaseName *uint16, desiredAcces ServiceSCMAccessRight) (handle DatabaseHandle, err error) {
+func _OpenSCManager(machineName *uint16, databaseName *uint16, desiredAcces ServiceSCMAccessRight) (handle Handle, err error) {
 	r0, _, e1 := syscall.Syscall(procOpenSCManagerW.Addr(), 3, uintptr(unsafe.Pointer(machineName)), uintptr(unsafe.Pointer(databaseName)), uintptr(desiredAcces))
-	handle = DatabaseHandle(r0)
+	handle = Handle(r0)
 	if handle == 0 {
 		if e1 != 0 {
 			err = errnoErr(e1)
@@ -77,7 +77,7 @@ func _OpenSCManager(machineName *uint16, databaseName *uint16, desiredAcces Serv
 	return
 }
 
-func _EnumServicesStatusEx(handle DatabaseHandle, infoLevel ServiceInfoLevel, serviceType ServiceType, serviceState ServiceEnumState, services *byte, bufSize uint32, bytesNeeded *uint32, servicesReturned *uint32, resumeHandle *uintptr, groupName *uintptr) (err error) {
+func _EnumServicesStatusEx(handle Handle, infoLevel ServiceInfoLevel, serviceType ServiceType, serviceState ServiceEnumState, services *byte, bufSize uint32, bytesNeeded *uint32, servicesReturned *uint32, resumeHandle *uintptr, groupName *uintptr) (err error) {
 	r1, _, e1 := syscall.Syscall12(procEnumServicesStatusExW.Addr(), 10, uintptr(handle), uintptr(infoLevel), uintptr(serviceType), uintptr(serviceState), uintptr(unsafe.Pointer(services)), uintptr(bufSize), uintptr(unsafe.Pointer(bytesNeeded)), uintptr(unsafe.Pointer(servicesReturned)), uintptr(unsafe.Pointer(resumeHandle)), uintptr(unsafe.Pointer(groupName)), 0, 0)
 	if r1 == 0 {
 		if e1 != 0 {
@@ -89,7 +89,7 @@ func _EnumServicesStatusEx(handle DatabaseHandle, infoLevel ServiceInfoLevel, se
 	return
 }
 
-func _OpenService(handle DatabaseHandle, serviceName *uint16, desiredAccess ServiceAccessRight) (serviceHandle Handle, err error) {
+func _OpenService(handle Handle, serviceName *uint16, desiredAccess ServiceAccessRight) (serviceHandle Handle, err error) {
 	r0, _, e1 := syscall.Syscall(procOpenServiceW.Addr(), 3, uintptr(handle), uintptr(unsafe.Pointer(serviceName)), uintptr(desiredAccess))
 	serviceHandle = Handle(r0)
 	if serviceHandle == 0 {
